feat: deploy test and prod stacks to uncefact AWS account
