+++ .github/workflows/ci_cd-untp-playground.yml
@@ -20,8 +20,7 @@ jobs:
     name: Deploy to Test
     environment:
       name: test
-      url: https://test-playground.untp.showthething.com/untp-playground
-
+      url: https://test.uncefact.org/test-untp-playground
     env:
       STACK_NAME: test
 
@@ -49,10 +48,9 @@ jobs:
           stack-name: ${{ env.STACK_NAME}}
           work-dir: ./packages/untp-playground/infra
         env:
-          PULUMI_CONFIG_PASSPHRASE: ${{ secrets.PULUMI_CONFIG_PASSPHRASE }}
-          NEXT_PUBLIC_BASE_PATH: /untp-playground
-          NEXT_PUBLIC_ASSET_PREFIX: /untp-playground
-          NEXT_PUBLIC_IMAGE_PATH: /untp-playground/_next/image
+          NEXT_PUBLIC_BASE_PATH: /test-untp-playground
+          NEXT_PUBLIC_ASSET_PREFIX: /test-untp-playground
+          NEXT_PUBLIC_IMAGE_PATH: /test-untp-playground/_next/image
 
   deploy_prod:
     if: github.repository_owner == 'uncefact' && github.ref_type == 'tag' && github.event_name == 'workflow_dispatch'
@@ -63,7 +61,7 @@ jobs:
       id-token: write
       contents: read
 
-    name: Deploy to Test
+    name: Deploy to Prod
     environment:
       name: production
       url: https://test.uncefact.org/untp-playground
@@ -95,7 +93,6 @@ jobs:
           stack-name: ${{ env.STACK_NAME}}
           work-dir: ./packages/untp-playground/infra
         env:
-          PULUMI_CONFIG_PASSPHRASE: ${{ secrets.PULUMI_CONFIG_PASSPHRASE }}
           NEXT_PUBLIC_BASE_PATH: /untp-playground
           NEXT_PUBLIC_ASSET_PREFIX: /untp-playground
           NEXT_PUBLIC_IMAGE_PATH: /untp-playground/_next/image
