+++ packages/untp-playground/README.md
@@ -6,23 +6,27 @@
 yarn dev
 ```
 
-Open [http://localhost:3000/untp-playground](http://localhost:3000/untp-playground) with your browser to see the result.
+Open [http://localhost:3000/](http://localhost:3000/) with your browser to see the result.
 
 ## Deployment
 
-We use Pulumi and GitHub actions to deploy the app. Please note that basePath is set to `/untp-playground`
+We use Pulumi and GitHub actions to deploy the app. The following env variables needs to be set to define basePath in order to it work with the existing CloudFront Distribution:
+```
+NEXT_PUBLIC_BASE_PATH: /untp-playground
+NEXT_PUBLIC_ASSET_PREFIX: /untp-playground
+NEXT_PUBLIC_IMAGE_PATH: /untp-playground/_next/image
+```
 
 GitHub cicd workflow requires the following secrets:
 
 1. PULUMI_AWS_SECRET_KEY_ID
 1. PULUMI_AWS_SECRET_ACCESS_KEY
-1. PULUMI_CONFIG_PASSPHRASE
 
-AWS credentials will be replaced with OIDC role in AWS account, Pulumi config encryption will be changed to awskms.
+AWS credentials will be replaced with OIDC role in AWS account, Pulumi config encryption has be changed to awskms.
 
 End-points:
 
-1. test - https://test-playground.untp.showthething.com/untp-playground
+1. test - https://test.uncefact.org/test-untp-playground
 1. production - https://test.uncefact.org/untp-playground
 
 `next` branch is getting automatically deployed to test, tag is manually deployed to production.
