+++ packages/untp-playground/infra/Pulumi.test.yaml
@@ -1,3 +1,4 @@
-encryptionsalt: v1:K20uwSxHGr8=:v1:+CqwmCEMvMU+56x4:P3zV3Azohm0se7uMdtnrd3Qlr/tbuQ==
+secretsprovider: awskms://alias/untp-pg-base?region=us-east-1
+encryptedkey: AQICAHiQPtPpMskboX6hzwJRSoxN1ESI7Hg9WiuyfnhBuWLLrAFxHhmNJ+lmjYpTAl3664O0AAAAfjB8BgkqhkiG9w0BBwagbzBtAgEAMGgGCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMr1XsDGUO1y8LDkCEAgEQgDt2zh99T15QUYnynakVTBrJEAKRpniRKBpV2XmI4sasMA4mcRjW15kuyzDmiOqMRPB2n1e1PMX6nPYtKA==
 config:
-  domain: "test-playground.untp.showthething.com"
+  appURL: "https://test.uncefact.org/test-untp-playground"
