+++ packages/untp-playground/infra/infra/index.ts
@@ -1,0 +1,6 @@
+import * as aws from "@pulumi/aws";
+import * as awsx from "@pulumi/awsx";
+
+export * from './base';
+export * from './state';
+export * from './app';
