+++ packages/untp-playground/infra/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "untp-demo",
+    "name": "untp-playground",
     "main": "index.ts",
     "devDependencies": {
         "@types/node": "^18",
